[build] Don't include GraphicsContextGL.h in WebCoreArgumentCoders.h https://bugs.webkit.org/show_bug.cgi?id=251238 rdar://problem/104722312
Reviewed by Dean Jackson.

WebCoreArgumentCoders only needs a few simple declarations from
GraphicsContextGL.h. Refactor GraphicsContextGL.h to extra enum
SimulatedEventForTesting and struct GraphicsContextGLActiveInfo, allowing the
required pieces to be included with out need to pull in the entire
expensive-to-parse header.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/graphics/GraphicsContextGL.h:
* Source/WebCore/platform/graphics/GraphicsContextGLActiveInfo.h: Added.
* Source/WebCore/platform/graphics/GraphicsContextGLEnums.h: Added.
* Source/WebKit/Shared/WebCoreArgumentCoders.h:

Canonical link: https://commits.webkit.org/259605@main
